Remove scripts and update init
I've removed the brew install scripts. They are no longer needed with the
use of the Brewfile. I've also edited the init.sh script to use the
brew bundle command. This command is the one that used the new Brewfile
to install software via brew.

I also made some updates to comments and promts.